refactor: move test helpers to src/__tests__ directory
Move test utilities from src/test/ to src/__tests__/ to follow
conventional test directory naming:
- api-test-helpers.ts -> src/__tests__/api-test-helpers.ts
- setup.ts -> src/__tests__/setup.ts

Update all import paths across 10 test files to reference the new
location. This aligns with the project's __tests__ directory
convention used throughout the codebase.

Related to #1411
