+++ turbo/apps/web/app/api/agent/runs/[id]/events/__tests__/route.test.ts
@@ -25,7 +25,7 @@ import {
   createDefaultComposeConfig,
   createTestCliToken,
   deleteTestCliToken,
-} from "../../../../../../../src/test/api-test-helpers";
+} from "../../../../../../../src/__tests__/api-test-helpers";
 
 // Mock Next.js headers() function
 vi.mock("next/headers", () => ({
