+++ turbo/apps/web/app/api/webhooks/agent/checkpoints/__tests__/route.test.ts
@@ -23,7 +23,7 @@ import {
   createTestRequest,
   createDefaultComposeConfig,
   createTestSandboxToken,
-} from "../../../../../../src/test/api-test-helpers";
+} from "../../../../../../src/__tests__/api-test-helpers";
 
 // Mock Next.js headers() function
 vi.mock("next/headers", () => ({
