+++ turbo/apps/web/app/api/webhooks/agent/heartbeat/__tests__/route.test.ts
@@ -11,7 +11,7 @@ import {
   createTestRequest,
   createDefaultComposeConfig,
   createTestSandboxToken,
-} from "../../../../../../src/test/api-test-helpers";
+} from "../../../../../../src/__tests__/api-test-helpers";
 
 // Mock Next.js headers() function
 vi.mock("next/headers", () => ({
