+++ turbo/apps/web/vitest.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "node",
-    setupFiles: "./src/test/setup.ts",
+    setupFiles: "./src/__tests__/setup.ts",
     // Don't override env vars, let them pass through from system
     // Run tests sequentially to avoid database race conditions
     fileParallelism: false,
